requirements.txt
# Core
openai>=1.12.0
python-dotenv>=1.0.1
rich>=13.7.0

# Human API Server (serve.py)
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0

# Channel integrations (install only what you need)
discord.py>=2.3.0          # For --channel discord
python-telegram-bot>=20.0  # For --channel telegram
twilio>=8.0.0              # For --channel whatsapp
flask>=3.0.0               # For --channel whatsapp (inbound webhook)
